requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
absl-py==2.0.0
asttokens==2.4.1
cachetools==5.3.2
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.1
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.43.1
fsspec==2023.4.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
grpcio==1.59.2
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
kiwisolver==1.4.5
Markdown==3.5.1
MarkupSafe==2.1.2
matplotlib==3.8.1
matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.0
numpy==1.26.1
oauthlib==3.2.2
onnx==1.15.0
onnxruntime-gpu==1.16.1
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.2
parso==0.8.3
Pillow==9.3.0
prompt-toolkit==3.0.39
protobuf==4.21.2
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygame==2.5.2
Pygments==2.16.1
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.0
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
stack-data==0.6.3
sympy==1.12
tensorboard==2.15.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
tokenizers==0.14.1
torch==2.1.0+cu118
torchinfo==1.8.0
torchvision==0.16.0+cu118
tqdm==4.66.1
traitlets==5.13.0
transformers==4.34.1
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
wcwidth==0.2.9
Werkzeug==3.0.1
yolov7detect==1.0.1
